requirements.txt
torch>=1.13.0
torchvision
transformers
numpy
Pillow
tqdm